Implement build_clique_collection()
This function uses networkx to find cliques in the mutual follower
graph, then stores those cliques in a clique collection. The
schema of documents in the clique collection include:

- `members`: a list of `str_ids` of members of the clique
- `size`: the number of members in the clique

The clique collection is rebuilt everytime `build_clique_collection`
is called.

This is inspired by Example 4-11 of *Mining the Social Web*.