Implmement build --cliques command
This provides command line access for the clique collection
construction pipeline with `networkx` and the `network.py` module
in particular.

The `--clique-cname` flag sets the collection name to store
clique documents in.

Added `networkx` as a dependency in `setup.py`.

Updated README with clique info; this need to be bolstered with
a tutorial.

Next steps: analyze cliques and make clique data available to
the user.